lmb-1453: move merge controller to the service
